server/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
groq
google-genai
cloudinary
pytest
httpx
python-jose[cryptography]
passlib[argon2]
email-validator
argon2-cffi
pypdf
supabase
mysql-connector-python
pymysql
